api-exemplo/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
PyJWT